\name{hex.spp}
\alias{hex.spp}
\title{Presence/Absence of Bird Species in Oregon, USA}
\description{
  Binary matrix (1 = present) for distributions of 248 native
  breeding bird species for 404 grid cells in Oregon, USA.
}
\usage{data (hex.spp)}
\format{A matrix with 404 rows and 248 columns.}
\details{
  Row labels are hexagon identifiers from White et al. (1992).
  Column labels are species element codes developed by 
  The Nature Conservancy (TNC), 
  the Oregon Natural Heritage Program (ONHP), and
  the Association for Biodiversity Information (ABI).
}
\source{Denis White, \email{white.denis@epa.gov} }
\references{
  Master, L. (1996) Predicting distributions for vertebrate 
  species: some observations, \emph{Gap Analysis: A Landscape 
  Approach to Biodiversity Planning}, Scott, J.M., Tear, T.H., 
  and Davis, F.W., editors, American Society for Photogrammetry 
  and Remote Sensing, Bethesda, MD, pp. 171-176.

  White, D., Preston, E.M., Freemark, K.E., Kiester, A.R. (1999) 
  A hierarchical framework for conserving biodiversity, 
  \emph{Landscape ecological analysis: issues and applications}, 
  Klopatek, J.M., Gardner, R.H., editors, Springer-Verlag, pp. 
  127-153.

  White, D., Kimerling, A.J., Overton, W.S.  (1992)  Cartographic  
  and geometric components of a global sampling design for 
  environmental monitoring, \emph{Cartography and Geographic 
  Information Systems}, \bold{19}(1), 5-22.

  TNC, \url{http://www.tnc.org/}

  ONHP, \url{http://www.heritage.tnc.org/nhp/us/or/}

  ABI, \url{http://www.abi.org/}

  TNC/ABI species data, \url{http://consci.tnc.org/src/overview.htm}
}
\seealso{
  \code{\link{hex.env}}, 
  \code{\link{spp.names}}, 
  \code{\link{oregon}}
}
\keyword{datasets}
