\name{oregon.bird.names}
\alias{spp.names}
\title{Names of Bird Species in Oregon, USA}
\description{
  Scientific and common names for 248 native breeding bird species
  in Oregon, USA.
}
\usage{data (oregon.bird.names)}
\format{A data frame with 248 rows and 2 columns.}
\details{
  Row names are species element codes.  Columns are
  \code{"scientific.name"} and \code{"common.name"}. 
  Data are provided by The Nature Conservancy (TNC), 
  the Oregon Natural Heritage Program (ONHP), and
  the Association for Biodiversity Information (ABI).
}
\source{Denis White, \email{white.denis@epa.gov} }
\references{
  Master, L. (1996) Predicting distributions for vertebrate 
  species: some observations, \emph{Gap Analysis: A Landscape 
  Approach to Biodiversity Planning}, Scott, J.M., Tear, T.H., 
  and Davis, F.W., editors, American Society for Photogrammetry 
  and Remote Sensing, Bethesda, MD, pp. 171-176.

  TNC, \url{http://nature.org/}

  ONHP, \url{http://www.abi.org/nhp/us/or/}

  ABI, \url{http://www.abi.org/}

  ABI species data, \url{http://www.abi.org/datasets_zoo/overview.htm}
}
\seealso{
  \code{\link{oregon.bird.dist}}
}
\keyword{datasets}
