\name{oregon.border}
\alias{oregon}
\title{Boundary of Oregon, USA}
\description{
  The boundary of the state of Oregon, USA, in 
  \code{\link{lines}} format.
}
\usage{data (oregon.border)}
\format{A data frame with 485 rows and 2 columns (the components 
  \code{"x"} and \code{"y"}).}
\details{
  The map projection for this boundary, as well as the point
  coordinates in \code{\link{oregon.env.vars}}, is the Lambert
  Conformal Conic with standard parallels at 33 and 45
  degrees North latitude, with the longitude of the central
  meridian at 120 degrees, 30 minutes West longitude,
  and with the projection origin latitude at 41 degrees,
  45 minutes North latitude.
}
\source{Denis White, \email{white.denis@epa.gov} }
\keyword{datasets}
