\name{spp.names}
\alias{spp.names}
\title{Names of Bird Species in Oregon, USA}
\description{
  Scientific and common names for 248 native breeding bird species
  in Oregon, USA.
}
\usage{data (spp.names)}
\format{A data frame with 248 rows and 2 columns.}
\details{
  Row labels are species element codes.  These and the
  names are provided by The Nature Conservancy (TNC), 
  the Oregon Natural Heritage Program (ONHP), and
  the Association for Biodiversity Information (ABI).
}
\source{Denis White, \email{white.denis@epa.gov} }
\references{
  Master, L. (1996) Predicting distributions for vertebrate 
  species: some observations, \emph{Gap Analysis: A Landscape 
  Approach to Biodiversity Planning}, Scott, J.M., Tear, T.H., 
  and Davis, F.W., editors, American Society for Photogrammetry 
  and Remote Sensing, Bethesda, MD, pp. 171-176.

  TNC, \url{http://www.tnc.org/}

  ONHP, \url{http://www.heritage.tnc.org/nhp/us/or/}

  ABI, \url{http://www.abi.org/}

  TNC/ABI species data, \url{http://consci.tnc.org/src/overview.htm}
}
\seealso{
  \code{\link{hex.spp}}
}
\keyword{datasets}
